+++ openequivariance/extension/generic_module.cpp
@@ -1,22 +1,26 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
+#include <iostream>
+#include <unordered_map>
+#include <stdexcept>
 
 #ifdef CUDA_BACKEND
-#include "backend_cuda.hpp"
-#include "group_mm_cuda.hpp"
-using JITKernel = CUJITKernel;
-using Allocator = CUDA_Allocator;
+    #include "backend_cuda.hpp"
+    #include "group_mm_cuda.hpp"
+    using JITKernel = CUJITKernel;
+    using GPU_Allocator = CUDA_Allocator;
 
-template<typename T>
-using GroupMM = GroupMMCUDA<T>; 
+    template<typename T>
+    using GroupMM = GroupMMCUDA<T>; 
 #endif
 
 #ifdef HIP_BACKEND
-#include "backend_hip.hpp"
-using JITKernel = HIPJITKernel;
-using Allocator = HIP_Allocator;
-template<typename T>
-using GroupMM = GroupMMHIP<T>; 
+    #include "backend_hip.hpp"
+    using JITKernel = HIPJITKernel;
+    using GPU_Allocator = HIP_Allocator;
+
+    template<typename T>
+    using GroupMM = GroupMMHIP<T>; 
 #endif
 
 #include "buffer.hpp"
@@ -24,15 +28,18 @@ using GroupMM = GroupMMHIP<T>;
 #include "convolution.hpp"
 
 using namespace std;
-namespace py = pybind11;
+namespace py=pybind11;
 
-PYBIND11_MODULE(kernel_wrapper, m) {
+PYBIND11_MODULE(generic_module, m) {
     //=========== Batch tensor products =========
     py::class_<GenericTensorProductImpl>(m, "GenericTensorProductImpl")
-        .def("exec_tensor_product", &GenericTensorProductImpl::exec_tensor_product_device_rawptrs)
-        .def("backward", &GenericTensorProductImpl::backward_device_rawptrs);
+        .def("exec_tensor_product_rawptr", &GenericTensorProductImpl::exec_tensor_product_device_rawptrs)
+        .def("backward_rawptr", &GenericTensorProductImpl::backward_device_rawptrs);
     py::class_<JITTPImpl<JITKernel>, GenericTensorProductImpl>(m, "JITTPImpl")
-        .def(py::init<std::string, KernelLaunchConfig, KernelLaunchConfig>());
+        .def(py::init<  std::string, 
+                        std::unordered_map<string, int64_t>, 
+                        std::unordered_map<string, int64_t>, 
+                        std::unordered_map<string, int64_t>>());
 
     //============= Convolutions ===============
     py::class_<ConvolutionImpl>(m, "ConvolutionImpl")
@@ -41,13 +48,12 @@ PYBIND11_MODULE(kernel_wrapper, m) {
     py::class_<JITConvImpl<JITKernel>, ConvolutionImpl>(m, "JITConvImpl")
         .def(py::init<std::string, KernelLaunchConfig, KernelLaunchConfig>());
 
-    //============= Utilities ===============
-    py::class_<KernelLaunchConfig>(m, "KernelLaunchConfig")
-        .def(py::init<>())
-        .def_readwrite("num_blocks", &KernelLaunchConfig::num_blocks)
-        .def_readwrite("warp_size", &KernelLaunchConfig::warp_size)
-        .def_readwrite("num_threads", &KernelLaunchConfig::num_threads)
-        .def_readwrite("smem", &KernelLaunchConfig::smem);
+    py::class_<GroupMM<float>>(m, "GroupMM_F32")
+        .def(py::init<int, int>())
+        .def("group_gemm", &GroupMM<float>::group_gemm_intptr);
+    py::class_<GroupMM<double>>(m, "GroupMM_F64")
+        .def(py::init<int, int>())
+        .def("group_gemm", &GroupMM<double>::group_gemm_intptr);
 
     py::class_<DeviceProp>(m, "DeviceProp")
         .def(py::init<int>())
@@ -58,22 +64,15 @@ PYBIND11_MODULE(kernel_wrapper, m) {
         .def_readonly("multiprocessorCount", &DeviceProp::multiprocessorCount)
         .def_readonly("maxSharedMemPerBlock", &DeviceProp::maxSharedMemPerBlock); 
 
-    py::class_<PyDeviceBuffer<Allocator>>(m, "DeviceBuffer")
+    py::class_<PyDeviceBuffer<GPU_Allocator>>(m, "DeviceBuffer")
         .def(py::init<uint64_t>())
         .def(py::init<py::buffer>())
-        .def("copy_to_host", &PyDeviceBuffer<Allocator>::copy_to_host)
-        .def("data_ptr", &PyDeviceBuffer<Allocator>::data_ptr);
+        .def("copy_to_host", &PyDeviceBuffer<GPU_Allocator>::copy_to_host)
+        .def("data_ptr", &PyDeviceBuffer<GPU_Allocator>::data_ptr);
 
     py::class_<GPUTimer>(m, "GPUTimer")
         .def(py::init<>())
         .def("start", &GPUTimer::start)
         .def("stop_clock_get_elapsed", &GPUTimer::stop_clock_get_elapsed)
         .def("clear_L2_cache", &GPUTimer::clear_L2_cache);
-
-    py::class_<GroupMM<float>>(m, "GroupMM_F32")
-        .def(py::init<int, int>())
-        .def("group_gemm", &GroupMM<float>::group_gemm_intptr);
-    py::class_<GroupMM<double>>(m, "GroupMM_F64")
-        .def(py::init<int, int>())
-        .def("group_gemm", &GroupMM<double>::group_gemm_intptr);
 }
