+++ openequivariance/extlib/__init__.py
@@ -1,4 +1,4 @@
-import os, warnings, tempfile
+import os, warnings, tempfile, warnings
 from pathlib import Path
 
 from openequivariance.benchmark.logging_utils import getLogger
@@ -6,18 +6,12 @@ from openequivariance.benchmark.logging_utils import getLogger
 oeq_root = str(Path(__file__).parent.parent)
 
 build_ext = True 
-candidates = [f for f in os.listdir(oeq_root + '/extlib') 
-                if f.startswith('torch_wrapper')]
-
-if len(candidates) == 1:
-    build_ext = False 
-
-torch_wrapper = None
-postprocess_kernel = lambda kernel: kernel
 TORCH_COMPILE=True
+torch_module, generic_module = None, None
+postprocess_kernel = lambda kernel: kernel
 
 if not build_ext: 
-    from openequivariance.extlib.torch_wrapper import * 
+    from openequivariance.extlib.generic_module import * 
 else:
     from setuptools import setup
     from torch.utils.cpp_extension import library_paths, include_paths
@@ -25,8 +19,9 @@ else:
     global torch
     import torch
 
-    sources = ['torch_wrapper.cpp']
     extra_cflags=["-O3"]
+    generic_sources = ['generic_module.cpp']
+    torch_sources = ['torch_tp_jit.cpp']
 
     include_dirs, extra_link_args = ['util'], None 
     if torch.cuda.is_available() and torch.version.cuda:
@@ -52,22 +47,29 @@ else:
 
         extra_cflags.append("-DHIP_BACKEND")
 
-    aten_sources = [oeq_root + '/extension/' + src for src in sources]
+    generic_sources = [oeq_root + '/extension/' + src for src in generic_sources]
+    torch_sources = [oeq_root + '/extension/' + src for src in torch_sources]
     include_dirs = [oeq_root + '/extension/' + d for d in include_dirs] + include_paths('cuda')
 
+    torch_compile_exception = None
     with warnings.catch_warnings():
-        #with tempfile.TemporaryDirectory() as tmpdir: // Might be needed for RocM
         warnings.simplefilter("ignore")
-        # If compiling torch fails (e.g. low gcc version), we should fall back to the
-        # version that takes integer pointers as args (but is untraceable to PyTorch JIT / export). 
-        try:
-            torch_wrapper = torch.utils.cpp_extension.load("torch_wrapper",
-                aten_sources, extra_cflags=extra_cflags + ["-DCOMPILE_TORCH"], extra_include_paths=include_dirs, extra_ldflags=extra_link_args)
-            torch.ops.load_library(torch_wrapper.__file__)
-        except Exception as e:
-            getLogger().warning("Warning, could not compile integrated PyTorch wrapper. Falling back to Pybind11" +
-                              ", but JITScript, compile fullgraph, and export will fail.")
-            torch_wrapper = torch.utils.cpp_extension.load("torch_wrapper",
-                aten_sources, extra_cflags=extra_cflags, extra_include_paths=include_dirs, extra_ldflags=extra_link_args)
 
-from torch_wrapper import *
+        try:
+            torch_module = torch.utils.cpp_extension.load("torch_tp_jit",
+                torch_sources, extra_cflags=extra_cflags, extra_include_paths=include_dirs, extra_ldflags=extra_link_args)
+            torch.ops.load_library(torch_module.__file__)
+        except Exception as e:
+            # If compiling torch fails (e.g. low gcc version), we should fall back to the
+            # version that takes integer pointers as args (but is untraceable to PyTorch JIT / export). 
+            TORCH_COMPILE=False
+            torch_compile_exception = e
+
+        generic_module = torch.utils.cpp_extension.load("generic_module",
+            generic_sources, extra_cflags=extra_cflags, extra_include_paths=include_dirs, extra_ldflags=extra_link_args)
+
+    if not TORCH_COMPILE:
+        warnings.warn("Could not compile integrated PyTorch wrapper. Falling back to Pybind11" +
+                            ", but JITScript, compile fullgraph, and export will fail.")
+
+from generic_module import *
