+++ openequivariance/implementations/TensorProduct.py
@@ -16,4 +16,4 @@ class TensorProduct(torch.nn.Module, LoopUnrollTP):
         return LoopUnrollTP.name()
 
     def forward(self, L1: torch.Tensor, L2: torch.Tensor, W: torch.Tensor) -> torch.Tensor:
-        return torch.ops.torch_wrapper.jit_tp_forward(self.internal, L1, L2, W)
+        return torch.ops.torch_module.jit_tp_forward(self.internal, L1, L2, W)
